+++ amazonka-s3-encryption/src/Network/AWS/S3/Encryption/Decrypt.hs
@@ -9,7 +9,6 @@ module Network.AWS.S3.Encryption.Decrypt where
 
 import Control.Lens ((%~), (^.))
 import qualified Control.Monad.Except as Except
-import Data.Coerce (coerce)
 import qualified Network.AWS as AWS
 import Network.AWS.Core
 import Network.AWS.Prelude